crewai_demo/requirements.txt
crewai
langchain-google-genai
litellm
crewai-tools
langchain-google-community[gmail]
